moved extensions to utils
