+++ LibrarySystem/LibrarySystem/Utils/SessionExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json;
 
-namespace LibrarySystem.Infrastructure;
+namespace LibrarySystem.Utils;
 
 public static class SessionExtensions
 {
@@ -12,6 +12,6 @@ public static class SessionExtensions
     public static T GetObjectFromJson<T>(this ISession session, string key)
     {
         var value = session.GetString(key);
-        return value == null ? default(T) : JsonConvert.DeserializeObject<T>(value);
+        return value == null ? default : JsonConvert.DeserializeObject<T>(value);
     }
 }
